Update NewEncoderCounter.cpp
Removed the direction variables because they don't add anything to the code. Note that this program does NOT allocate the A5 and A6 pins for power to the Encoders, but it can be set up that way if you want. The Vcc pins are instead hooked directly up to the 3.3 V and 5V pins.